% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBMA_initial_python_setup.R
\name{clusterBMA_initial_python_setup}
\alias{clusterBMA_initial_python_setup}
\title{Set up Conda environment using `miniconda` with Python and Tensorflow (Run once on first install)}
\usage{
clusterBMA_initial_python_setup()
}
\value{
Sets up conda environment with Python and Tensorflow
}
\description{
Set up Conda environment using `miniconda` with Python and Tensorflow (Run once on first install)
}
