% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBMA_use_condaenv.R
\name{clusterBMA_use_condaenv}
\alias{clusterBMA_use_condaenv}
\title{Point reticulate and tensorflow to the correct conda environment}
\usage{
clusterBMA_use_condaenv()
}
\description{
Point reticulate and tensorflow to the correct conda environment
}
\examples{
Run at start when setting up a session for clusterBMA: clusterBMA_use_condaenv()
}
