% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_matrix_fn.R
\name{consensus_matrix_fn}
\alias{consensus_matrix_fn}
\title{Generate consensus matrix from input similarity matrices, and calculate key outputs (internal function)}
\usage{
consensus_matrix_fn(sim_mat_list, algo_weights, n_final_clust, prior_vec)
}
\description{
Generate consensus matrix from input similarity matrices, and calculate key outputs (internal function)
}
