% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_to_prob_fn.R
\name{hard_to_prob_fn}
\alias{hard_to_prob_fn}
\title{Turn cluster labels from 'hard' clustering algorithm into an allocation probability matrix}
\usage{
hard_to_prob_fn(hard_clusters, n_clust)
}
\arguments{
\item{hard_clusters}{A vector of cluster allocations from a hard clustering algorithm of length N (number of observations in clustered dataset)}

\item{n_clust}{The number of clusters, k, in the clustering solution}
}
\value{
Returns a N x k matrix of cluster allocation probabilities, which is used as an input for `clusterBMA::clusterBMA()`
}
\description{
`hard_to_prob_fn()` is used to generate a matrix of probabilities (1 or 0) of allocation to clusters, based on a list of cluster labels generated by a 'hard' clustering algorithm (binary/non-probabilistic allocation)
}
