% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_weight_fn.R
\name{new_weight_fn}
\alias{new_weight_fn}
\title{Calculate model weights based on Cluster Internal Validation Index (internal function)}
\usage{
new_weight_fn(
  input_data,
  cluster_label_df,
  n_sols,
  wt_crit_name,
  wt_crit_direction
)
}
\description{
Calculate model weights based on Cluster Internal Validation Index (internal function)
}
