% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_to_hard_fn.R
\name{prob_to_hard_fn}
\alias{prob_to_hard_fn}
\title{(internal function) Generate 'hard' allocations and probability/uncertainty from soft allocation probability matrix ($P output from SimplexClust())}
\usage{
prob_to_hard_fn(soft_alloc_matrix)
}
\description{
(internal function) Generate 'hard' allocations and probability/uncertainty from soft allocation probability matrix ($P output from SimplexClust())
}
