% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mat_fn.R
\name{sim_mat_fn}
\alias{sim_mat_fn}
\title{(internal function) Generate similarity matrix from allocation probability matrix}
\usage{
sim_mat_fn(cluster_probs)
}
\description{
(internal function) Generate similarity matrix from allocation probability matrix
}
