% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexClust.R
\name{SimplexClust}
\alias{SimplexClust}
\title{(internal function) Symmetric Simplex Matrix Factorisation using Tensorflow
Generate model averaged clusters and allocations based on consensus matrix}
\usage{
SimplexClust(S, G, lambda = 1, eps = 0.01, steps = 5000, choice_loss = "KL")
}
\arguments{
\item{S}{an n x n consensus matrix}

\item{G}{number of clusters}

\item{lambda}{the amount of L2 regularization (1 by default)}

\item{steps}{number of iterations (more = marginally better convergence. Mostly seems to converge after ~2000 steps (original default))}
}
\description{
(internal function) Symmetric Simplex Matrix Factorisation using Tensorflow
Generate model averaged clusters and allocations based on consensus matrix
}
